% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{blending}
\alias{blending}
\title{Blending}
\usage{
blending(X, y, models, paramses, preprocesses = NULL,
  objective = c("regression", "classification"), test_prop = 0.1,
  train_prop = switch(length(models), `1` = 1, `2` = c(0.9, 0.1), `3` = c(0.7,
  0.2, 0.1)))
}
\arguments{
\item{X}{data.frame of independent variable}

\item{y}{vector of dependent variable}

\item{models}{list of layers, each layer is a list of blending model class}

\item{objective}{what kind of problem}

\item{test_prop}{what's the proportion of test data}

\item{train_prop}{what's the proportion of data for each layer training}

\item{params}{list of parameters for corresponding model}
}
\value{
a blending model
}
\description{
Blending
}

