% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{default_preprocesses}
\alias{default_preprocesses}
\title{Default preprocesses}
\usage{
default_preprocesses(models)
}
\arguments{
\item{model}{model list}
}
\value{
default preprocesses list
}
\description{
Default preprocesses
}

