% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{blending}
\alias{blending}
\title{Blending}
\usage{
blending(X, y, models, paramses, preprocesses = NULL,
  objective = c("regression", "classification"), test_prop = 0.1,
  train_prop = switch(length(models), `1` = 1, `2` = c(0.9, 0.1), `3` = c(0.7,
  0.2, 0.1)))
}
\arguments{
\item{X}{data.frame of independent variable}

\item{y}{vector of dependent variable}

\item{models}{list of layers, each layer is a list of blending model class}

\item{objective}{what kind of problem}

\item{test_prop}{what's the proportion of test data}

\item{train_prop}{what's the proportion of data for each layer training}

\item{params}{list of parameters for corresponding model}
}
\value{
a blending model
}
\description{
Blending
}
\examples{
# regression
models = list(
  list(Caret, XGBoost),
  list(XGBoost))

paramses = list(
  list(list(method = 'bridge'), list(nrounds = 50)),
  list(list(nrounds = 5)))

preprocesses = default_preprocesses(models)

X = iris[,1:3]
y = iris[,4]
m = blending(X, y, models, paramses, preprocesses, objective = "regression")
predict(m, X)

# classification
models = list(
  list(XGBoost, XGBoost),
  list(XGBoost))

paramses = list(
  list(list(nrounds = 50, objective = "multi:softmax", num_class = 3), list(nrounds = 10, objective = "multi:softmax", num_class = 3)),
  list(list(nrounds = 5, objective = "multi:softmax", num_class = 3)))

preprocesses = default_preprocesses(models)

X = iris[,1:4]
y = iris[,5]
m = blending(X, y, models, paramses, preprocesses, objective = "classification")
predict(m, X)
}

